requirements.txt
affine==2.4.0
albumentations==1.3.1
astroid==3.0.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dill==0.3.7
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.45.1
fsspec==2023.10.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.6
imageio==2.33.0
iniconfig==2.0.0
invoke==2.2.0
isort==5.12.0
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
mando==0.7.1
MarkupSafe==2.1.3
matplotlib==3.8.2
mccabe==0.7.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
pycodestyle==2.11.1
pylint==3.0.2
pyparsing==3.1.1
pytest==7.4.3
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
radon==6.0.1
rasterio==1.3.9
requests==2.31.0
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
smmap==5.0.1
snuggs==1.4.7
sympy==1.12
threadpoolctl==3.2.0
tifffile==2023.9.26
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
xenon==0.9.1
